Restructured and cleaned up files
