--- backend/src/main/java/tutraffic/backend/BackendApplication.java
@@ -1,12 +1,0 @@
-package tutraffic.backend;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class BackendApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(BackendApplication.class, args);
-	}
-}
