--- backend/src/main/java/tutraffic/backend/UserRepository.java
@@ -1,10 +1,0 @@
-package tutraffic.backend;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
-
-    User findByEmail(String email);
-}
