--- backend/src/main/resources/application.properties
@@ -1,11 +1,0 @@
-# The embedded database can be found in the user home directory in the "/tutraffic" folder
-spring.datasource.url=jdbc:h2:~/tutraffic/database
-spring.datasource.username=sa
-spring.datasource.password=
-spring.jpa.show-sql=true
-
-spring.jpa.hibernate.ddl-auto=update
-spring.jpa.properties.hibernate.format_sql=true
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.H2Dialect
-spring.h2.console.enabled=true
-spring.h2.console.settings.web-allow-others=true
