--- backend/src/test/java/tutraffic/backend/BackendApplicationTests.java
@@ -1,13 +1,0 @@
-package tutraffic.backend;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class BackendApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
